Windows Extensions: Fix potential AttributeErrors
If `self.get_owner()` returns `None`, and the chained call to
`is_valid()` is executed, an `AttributeError` will occur. This fixes two
instances of this bug by intializing a local variable with the result of
the `get_owner()` call, checking for `None`, and returning if that's the
case.

Also adds type-hints for these methods.